+++ Avia Terra Core/src/main/java/net/atcore/aviaterraplayer/ModerationPlayer.java
@@ -17,7 +17,7 @@ public class ModerationPlayer extends AbstractAviaTerraPlayer {
     }
 
     private int sanctionsChat = 1;//por circunstancias matemáticas tiene que ser 1
-    private float pointChat = ChatModeration.MAX_PUNTOS;
+    private double pointChat = ChatModeration.MAX_PUNTOS;
     private boolean isFreeze = false;
     private List<UUID> manipulatorInventoryPlayer = new ArrayList<>();
     private UUID manipulatedInventoryPlayer = null;
